Delay the fade-in animation until the fps calculation is done.
In some small devices the animation is causing that the fps calculation
does not get the correct number because the animation is too heavy.
This change adds some seconds to the start of the initial fade-in animation
to make sure it happens when the fps calculation is done. We also
delay the next color change animation accordingly.